+++ test/extensions_and_deriving/test.ml
@@ -51,7 +51,8 @@ val gen_type_decl : Context_free.Rule.t = <abstr>
 
 [%%expect{|
 Line _, characters 3-19:
-Error: Attributes not allowed here
+Error: The following located exception was raised during the context-free transformation phase:
+Attributes not allowed here
 |}]
 
 (* Generates a [type t = int[@@deriving derived]] *)